Combine all command line functionality into one
