+++ CMakeLists.txt
@@ -100,10 +100,8 @@ if(IS_MSVC)
     add_definitions("-DGLOG_NO_ABBREVIATED_SEVERITIES")
     add_definitions("-DGL_GLEXT_PROTOTYPES")
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /EHsc")
-    set(COLMAP_LIB_PATH "bin")
 else()
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
-    set(COLMAP_LIB_PATH "lib/colmap")
 endif()
 
 if(IS_GNU)
@@ -201,16 +199,10 @@ add_subdirectory(src)
 ################################################################################
 
 if(IS_MSVC)
-    install(FILES "scripts/shell/colmap.bat" "scripts/shell/run_tests.bat"
+    install(FILES "scripts/shell/COLMAP.bat" "scripts/shell/RUN_TESTS.bat"
             PERMISSIONS OWNER_READ OWNER_WRITE OWNER_EXECUTE
                         GROUP_READ GROUP_EXECUTE WORLD_READ WORLD_EXECUTE
             DESTINATION "/")
-else()
-    install(CODE "FILE(MAKE_DIRECTORY \${CMAKE_INSTALL_PREFIX}/lib/colmap)")
-    install(FILES "scripts/shell/colmap.sh"
-            PERMISSIONS OWNER_READ OWNER_WRITE OWNER_EXECUTE
-                        GROUP_READ GROUP_EXECUTE WORLD_READ WORLD_EXECUTE
-            DESTINATION "bin" RENAME "colmap")
 endif()
 
 configure_file("${CMAKE_SOURCE_DIR}/cmake/CMakeUninstall.cmake"
