+++ cmake/CMakeHelper.cmake
@@ -199,7 +199,7 @@ macro(COLMAP_ADD_EXECUTABLE TARGET_NAME)
     target_link_libraries(${TARGET_NAME} ${COLMAP_LIBRARIES})
     qt5_use_modules(${TARGET_NAME} ${COLMAP_QT_MODULES})
     COLMAP_ADD_TARGET_HELPER(${TARGET_NAME})
-    install(TARGETS ${TARGET_NAME} DESTINATION ${COLMAP_LIB_PATH})
+    install(TARGETS ${TARGET_NAME} DESTINATION bin/)
 endmacro(COLMAP_ADD_EXECUTABLE)
 
 # Wrapper for test executables
@@ -212,7 +212,9 @@ macro(COLMAP_ADD_TEST TARGET_NAME)
                               ${Boost_UNIT_TEST_FRAMEWORK_LIBRARY})
         COLMAP_ADD_TARGET_HELPER(${TARGET_NAME})
         add_test("${FOLDER_NAME}/${TARGET_NAME}" ${TARGET_NAME})
-        install(TARGETS ${TARGET_NAME} DESTINATION ${COLMAP_LIB_PATH})
+        if(IS_MSVC)
+            install(TARGETS ${TARGET_NAME} DESTINATION bin/)
+        endif()
     endif()
 endmacro(COLMAP_ADD_TEST)
 
@@ -226,6 +228,8 @@ macro(COLMAP_CUDA_ADD_TEST TARGET_NAME)
                               ${Boost_UNIT_TEST_FRAMEWORK_LIBRARY})
         COLMAP_ADD_TARGET_HELPER(${TARGET_NAME})
         add_test("${FOLDER_NAME}/${TARGET_NAME}" ${TARGET_NAME})
-        install(TARGETS ${TARGET_NAME} DESTINATION ${COLMAP_LIB_PATH})
+        if(IS_MSVC)
+            install(TARGETS ${TARGET_NAME} DESTINATION bin/)
+        endif()
     endif()
 endmacro(COLMAP_CUDA_ADD_TEST)
