+++ doc/gui.rst
@@ -3,13 +3,15 @@
 Graphical User Interface
 ========================
 
-The graphical user interface of COLMAP provides access to most of its
+The graphical user interface of COLMAP provides access to most of the available
 functionality and visualizes the reconstruction process in "real-time". To start
-the GUI, please run the pre-built binaries (Windows: `colmap.bat`, Mac:
-`colmap.app`), execute ``colmap gui`` if you installed COLMAP or execute
-``./src/exe/colmap`` from the build folder. The GUI application requires an
-attached display with OpenGL 3.2 support. Registered images are visualized in
-red and points in their average point color extracted from the images.
+the GUI, you can run the pre-built packages (Windows: `COLMAP.bat`, Mac:
+`COLMAP.app`), execute ``colmap gui`` if you installed COLMAP or execute
+``./src/exe/colmap gui`` from the CMake build folder. The GUI application
+requires an attached display with at least OpenGL 3.2 support. Registered images
+are visualized in red and reconstructed points in their average point color
+extracted from the images. The viewer can also visualize dense point clouds
+produced from Multi-View Stereo.
 
 
 Model Viewer Controls
