--- src/exe/bundle_adjuster.cc
@@ -1,46 +1,0 @@
-// COLMAP - Structure-from-Motion and Multi-View Stereo.
-// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include "base/reconstruction.h"
-#include "controllers/bundle_adjustment.h"
-#include "util/logging.h"
-#include "util/option_manager.h"
-
-using namespace colmap;
-
-int main(int argc, char** argv) {
-  InitializeGlog(argv);
-
-  std::string input_path;
-  std::string output_path;
-
-  OptionManager options;
-  options.AddRequiredOption("input_path", &input_path);
-  options.AddRequiredOption("output_path", &output_path);
-  options.AddBundleAdjustmentOptions();
-  options.Parse(argc, argv);
-
-  Reconstruction reconstruction;
-  reconstruction.Read(input_path);
-
-  BundleAdjustmentController ba_controller(options, &reconstruction);
-  ba_controller.Start();
-  ba_controller.Wait();
-
-  reconstruction.Write(output_path);
-
-  return EXIT_SUCCESS;
-}
