--- src/exe/color_extractor.cc
@@ -1,41 +1,0 @@
-// COLMAP - Structure-from-Motion and Multi-View Stereo.
-// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include "base/reconstruction.h"
-#include "util/logging.h"
-#include "util/option_manager.h"
-
-using namespace colmap;
-
-int main(int argc, char** argv) {
-  InitializeGlog(argv);
-
-  std::string import_path;
-  std::string export_path;
-
-  OptionManager options;
-  options.AddImageOptions();
-  options.AddDefaultOption("import_path", &import_path);
-  options.AddRequiredOption("export_path", &export_path);
-  options.Parse(argc, argv);
-
-  Reconstruction reconstruction;
-  reconstruction.Read(import_path);
-  reconstruction.ExtractColorsForAllImages(*options.image_path);
-  reconstruction.Write(export_path);
-
-  return EXIT_SUCCESS;
-}
