--- src/exe/exhaustive_matcher.cc
@@ -1,56 +1,0 @@
-// COLMAP - Structure-from-Motion and Multi-View Stereo.
-// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include <QApplication>
-
-#include "feature/matching.h"
-#include "util/logging.h"
-#include "util/option_manager.h"
-
-using namespace colmap;
-
-int main(int argc, char** argv) {
-  InitializeGlog(argv);
-
-#ifdef CUDA_ENABLED
-  const bool kUseOpenGL = false;
-#else
-  const bool kUseOpenGL = true;
-#endif
-
-  OptionManager options;
-  options.AddDatabaseOptions();
-  options.AddExhaustiveMatchingOptions();
-  options.Parse(argc, argv);
-
-  std::unique_ptr<QApplication> app;
-  if (options.sift_matching->use_gpu && kUseOpenGL) {
-    app.reset(new QApplication(argc, argv));
-  }
-
-  ExhaustiveFeatureMatcher feature_matcher(*options.exhaustive_matching,
-                                           *options.sift_matching,
-                                           *options.database_path);
-
-  if (options.sift_matching->use_gpu && kUseOpenGL) {
-    RunThreadWithOpenGLContext(&feature_matcher);
-  } else {
-    feature_matcher.Start();
-    feature_matcher.Wait();
-  }
-
-  return EXIT_SUCCESS;
-}
