--- src/exe/feature_importer.cc
@@ -1,66 +1,0 @@
-// COLMAP - Structure-from-Motion and Multi-View Stereo.
-// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include "base/camera_models.h"
-#include "feature/extraction.h"
-#include "util/logging.h"
-#include "util/misc.h"
-#include "util/option_manager.h"
-
-using namespace colmap;
-
-int main(int argc, char** argv) {
-  InitializeGlog(argv);
-
-  std::string import_path;
-  std::string image_list_path;
-
-  OptionManager options;
-  options.AddDatabaseOptions();
-  options.AddImageOptions();
-  options.AddRequiredOption("import_path", &import_path);
-  options.AddDefaultOption("image_list_path", &image_list_path);
-  options.AddExtractionOptions();
-  options.Parse(argc, argv);
-
-  ImageReaderOptions reader_options = *options.image_reader;
-  reader_options.database_path = *options.database_path;
-  reader_options.image_path = *options.image_path;
-
-  if (!image_list_path.empty()) {
-    reader_options.image_list = ReadTextFileLines(image_list_path);
-    if (reader_options.image_list.empty()) {
-      return EXIT_SUCCESS;
-    }
-  }
-
-  const std::vector<double> camera_params =
-      CSVToVector<double>(options.image_reader->camera_params);
-  const int camera_model_id =
-      CameraModelNameToId(options.image_reader->camera_model);
-
-  if (camera_params.size() > 0 &&
-      !CameraModelVerifyParams(camera_model_id, camera_params)) {
-    std::cerr << "ERROR: Invalid camera parameters" << std::endl;
-    return EXIT_FAILURE;
-  }
-
-  FeatureImporter feature_importer(reader_options, import_path);
-  feature_importer.Start();
-  feature_importer.Wait();
-
-  return EXIT_SUCCESS;
-}
