--- src/exe/mapper.cc
@@ -1,96 +1,0 @@
-// COLMAP - Structure-from-Motion and Multi-View Stereo.
-// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include "controllers/incremental_mapper.h"
-#include "util/logging.h"
-#include "util/misc.h"
-#include "util/option_manager.h"
-
-using namespace colmap;
-
-int main(int argc, char** argv) {
-  InitializeGlog(argv);
-
-  std::string import_path;
-  std::string export_path;
-  std::string image_list_path;
-
-  OptionManager options;
-  options.AddDatabaseOptions();
-  options.AddImageOptions();
-  options.AddDefaultOption("import_path", &import_path);
-  options.AddRequiredOption("export_path", &export_path);
-  options.AddDefaultOption("image_list_path", &image_list_path);
-  options.AddMapperOptions();
-  options.Parse(argc, argv);
-
-  if (!ExistsDir(export_path)) {
-    std::cerr << "ERROR: `export_path` is not a directory." << std::endl;
-    return EXIT_FAILURE;
-  }
-
-  if (!image_list_path.empty()) {
-    const auto image_names = ReadTextFileLines(image_list_path);
-    options.mapper->image_names =
-        std::set<std::string>(image_names.begin(), image_names.end());
-  }
-
-  ReconstructionManager reconstruction_manager;
-  if (import_path != "") {
-    if (!ExistsDir(import_path)) {
-      std::cerr << "ERROR: `import_path` is not a directory." << std::endl;
-      return EXIT_FAILURE;
-    }
-    reconstruction_manager.Read(import_path);
-  }
-
-  IncrementalMapperController mapper(options.mapper.get(), *options.image_path,
-                                     *options.database_path,
-                                     &reconstruction_manager);
-
-  // In case a new reconstruction is started, write results of individual sub-
-  // models to as their reconstruction finishes instead of writing all results
-  // after all reconstructions finished.
-  size_t prev_num_reconstructions = 0;
-  if (import_path == "") {
-    mapper.AddCallback(
-        IncrementalMapperController::LAST_IMAGE_REG_CALLBACK, [&]() {
-          // If the number of reconstructions has not changed, the last model
-          // was discarded for some reason.
-          if (reconstruction_manager.Size() > prev_num_reconstructions) {
-            const std::string reconstruction_path = JoinPaths(
-                export_path, std::to_string(prev_num_reconstructions));
-            const auto& reconstruction =
-                reconstruction_manager.Get(prev_num_reconstructions);
-            CreateDirIfNotExists(reconstruction_path);
-            reconstruction.Write(reconstruction_path);
-            options.Write(JoinPaths(reconstruction_path, "project.ini"));
-            prev_num_reconstructions = reconstruction_manager.Size();
-          }
-        });
-  }
-
-  mapper.Start();
-  mapper.Wait();
-
-  // In case the reconstruction is continued from an existing reconstruction, do
-  // not create sub-folders but directly write the results.
-  if (import_path != "" && reconstruction_manager.Size() > 0) {
-    reconstruction_manager.Get(0).Write(export_path);
-  }
-
-  return EXIT_SUCCESS;
-}
