--- src/exe/matches_importer.cc
@@ -1,75 +1,0 @@
-// COLMAP - Structure-from-Motion and Multi-View Stereo.
-// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include <QApplication>
-
-#include "feature/matching.h"
-#include "util/logging.h"
-#include "util/option_manager.h"
-
-using namespace colmap;
-
-int main(int argc, char** argv) {
-  InitializeGlog(argv);
-
-#ifdef CUDA_ENABLED
-  const bool kUseOpenGL = false;
-#else
-  const bool kUseOpenGL = true;
-#endif
-
-  std::string match_list_path;
-  std::string match_type = "pairs";
-
-  OptionManager options;
-  options.AddDatabaseOptions();
-  options.AddRequiredOption("match_list_path", &match_list_path);
-  options.AddDefaultOption("match_type", &match_type,
-                           "{'pairs', 'raw', 'inliers'}");
-  options.AddMatchingOptions();
-  options.Parse(argc, argv);
-
-  std::unique_ptr<QApplication> app;
-  if (options.sift_matching->use_gpu && kUseOpenGL) {
-    app.reset(new QApplication(argc, argv));
-  }
-
-  std::unique_ptr<Thread> feature_matcher;
-  if (match_type == "pairs") {
-    ImagePairsMatchingOptions matcher_options;
-    matcher_options.match_list_path = match_list_path;
-    feature_matcher.reset(new ImagePairsFeatureMatcher(
-        matcher_options, *options.sift_matching, *options.database_path));
-  } else if (match_type == "raw" || match_type == "inliers") {
-    FeaturePairsMatchingOptions matcher_options;
-    matcher_options.match_list_path = match_list_path;
-    matcher_options.verify_matches = match_type == "raw";
-    feature_matcher.reset(new FeaturePairsFeatureMatcher(
-        matcher_options, *options.sift_matching, *options.database_path));
-  } else {
-    std::cerr << "ERROR: Invalid `match_type`";
-    return EXIT_FAILURE;
-  }
-
-  if (options.sift_matching->use_gpu && kUseOpenGL) {
-    RunThreadWithOpenGLContext(feature_matcher.get());
-  } else {
-    feature_matcher->Start();
-    feature_matcher->Wait();
-  }
-
-  return EXIT_SUCCESS;
-}
