--- src/exe/point_triangulator.cc
@@ -1,148 +1,0 @@
-// COLMAP - Structure-from-Motion and Multi-View Stereo.
-// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include "controllers/incremental_mapper.h"
-#include "util/logging.h"
-#include "util/misc.h"
-#include "util/option_manager.h"
-
-using namespace colmap;
-
-int main(int argc, char** argv) {
-  InitializeGlog(argv);
-
-  std::string import_path;
-  std::string export_path;
-
-  OptionManager options;
-  options.AddDatabaseOptions();
-  options.AddImageOptions();
-  options.AddRequiredOption("import_path", &import_path);
-  options.AddRequiredOption("export_path", &export_path);
-  options.AddMapperOptions();
-  options.Parse(argc, argv);
-
-  if (!ExistsDir(import_path)) {
-    std::cerr << "ERROR: `import_path` is not a directory" << std::endl;
-    return EXIT_FAILURE;
-  }
-
-  if (!ExistsDir(export_path)) {
-    std::cerr << "ERROR: `export_path` is not a directory" << std::endl;
-    return EXIT_FAILURE;
-  }
-
-  const auto& mapper_options = *options.mapper;
-
-  PrintHeading1("Loading database");
-
-  DatabaseCache database_cache;
-
-  {
-    Database database(*options.database_path);
-    Timer timer;
-    timer.Start();
-    const size_t min_num_matches =
-        static_cast<size_t>(mapper_options.min_num_matches);
-    database_cache.Load(database, min_num_matches,
-                        mapper_options.ignore_watermarks,
-                        mapper_options.image_names);
-    std::cout << std::endl;
-    timer.PrintMinutes();
-  }
-
-  std::cout << std::endl;
-
-  Reconstruction reconstruction;
-  reconstruction.Read(import_path);
-
-  CHECK_GE(reconstruction.NumRegImages(), 2)
-      << "Need at least two images for triangulation";
-
-  IncrementalMapper mapper(&database_cache);
-  mapper.BeginReconstruction(&reconstruction);
-
-  //////////////////////////////////////////////////////////////////////////////
-  // Triangulation
-  //////////////////////////////////////////////////////////////////////////////
-
-  const auto tri_options = mapper_options.Triangulation();
-
-  for (const image_t image_id : reconstruction.RegImageIds()) {
-    const auto& image = reconstruction.Image(image_id);
-
-    PrintHeading1("Triangulating image #" + std::to_string(image_id));
-
-    const size_t num_existing_points3D = image.NumPoints3D();
-
-    std::cout << "  => Image has " << num_existing_points3D << " / "
-              << image.NumObservations() << " points" << std::endl;
-
-    mapper.TriangulateImage(tri_options, image_id);
-
-    std::cout << "  => Triangulated "
-              << (image.NumPoints3D() - num_existing_points3D) << " points"
-              << std::endl;
-  }
-
-  //////////////////////////////////////////////////////////////////////////////
-  // Bundle adjustment
-  //////////////////////////////////////////////////////////////////////////////
-
-  CompleteAndMergeTracks(mapper_options, &mapper);
-
-  const auto ba_options = mapper_options.GlobalBundleAdjustment();
-
-  // Configure bundle adjustment.
-  BundleAdjustmentConfig ba_config;
-  for (const image_t image_id : reconstruction.RegImageIds()) {
-    ba_config.AddImage(image_id);
-    ba_config.SetConstantPose(image_id);
-    ba_config.SetConstantCamera(reconstruction.Image(image_id).CameraId());
-  }
-
-  for (int i = 0; i < mapper_options.ba_global_max_refinements; ++i) {
-    // Avoid degeneracies in bundle adjustment.
-    reconstruction.FilterObservationsWithNegativeDepth();
-
-    const size_t num_observations = reconstruction.ComputeNumObservations();
-
-    PrintHeading1("Bundle adjustment");
-    BundleAdjuster bundle_adjuster(ba_options, ba_config);
-    CHECK(bundle_adjuster.Solve(&reconstruction));
-
-    size_t num_changed_observations = 0;
-    num_changed_observations += CompleteAndMergeTracks(mapper_options, &mapper);
-    num_changed_observations += FilterPoints(mapper_options, &mapper);
-    const double changed =
-        static_cast<double>(num_changed_observations) / num_observations;
-    std::cout << StringPrintf("  => Changed observations: %.6f", changed)
-              << std::endl;
-    if (changed < mapper_options.ba_global_max_refinement_change) {
-      break;
-    }
-  }
-
-  PrintHeading1("Extracting colors");
-  reconstruction.ExtractColorsForAllImages(*options.image_path);
-
-  const bool kDiscardReconstruction = false;
-  mapper.EndReconstruction(kDiscardReconstruction);
-
-  reconstruction.Write(export_path);
-
-  return EXIT_SUCCESS;
-}
