--- src/exe/rig_bundle_adjuster.cc
@@ -1,195 +1,0 @@
-// COLMAP - Structure-from-Motion and Multi-View Stereo.
-// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include "base/reconstruction.h"
-#include "optim/bundle_adjustment.h"
-#include "util/logging.h"
-#include "util/misc.h"
-#include "util/option_manager.h"
-
-#include <boost/property_tree/json_parser.hpp>
-#include <boost/property_tree/ptree.hpp>
-
-using namespace colmap;
-
-// Read the configuration of the camera rigs from a JSON file. The input images
-// of a camera rig must be named consistently to assign them to the appropriate
-// camera rig and the respective snapshots.
-//
-// An example configuration of a single camera rig:
-// [
-//   {
-//     "ref_camera_id": 1,
-//     "cameras":
-//     [
-//       {
-//           "camera_id": 1,
-//           "image_prefix": "left1_image"
-//       },
-//       {
-//           "camera_id": 2,
-//           "image_prefix": "left2_image"
-//       },
-//       {
-//           "camera_id": 3,
-//           "image_prefix": "right1_image"
-//       },
-//       {
-//           "camera_id": 4,
-//           "image_prefix": "right2_image"
-//       }
-//     ]
-//   }
-// ]
-//
-// This file specifies the configuration for a single camera rig and that you
-// could potentially define multiple camera rigs. The rig is composed of 4
-// cameras: all images of the first camera must have "left1_image" as a name
-// prefix, e.g., "left1_image_frame000.png" or "left1_image/frame000.png".
-// Images with the same suffix ("_frame000.png" and "/frame000.png") are
-// assigned to the same snapshot, i.e., they are assumed to be captured at the
-// same time. Only snapshots with the reference image registered will be added
-// to the bundle adjustment problem. The remaining images will be added with
-// independent poses to the bundle adjustment problem. The above configuration
-// could have the following input image file structure:
-//
-//    /path/to/images/...
-//        left1_image/...
-//            frame000.png
-//            frame001.png
-//            frame002.png
-//            ...
-//        left2_image/...
-//            frame000.png
-//            frame001.png
-//            frame002.png
-//            ...
-//        right1_image/...
-//            frame000.png
-//            frame001.png
-//            frame002.png
-//            ...
-//        right2_image/...
-//            frame000.png
-//            frame001.png
-//            frame002.png
-//            ...
-//
-// TODO: Provide an option to manually / explicitly set the relative extrinsics
-// of the camera rig. At the moment, the relative extrinsics are automatically
-// inferred from the reconstruction.
-std::vector<CameraRig> ReadCameraRigConfig(
-    const std::string& rig_config_path, const Reconstruction& reconstruction) {
-  boost::property_tree::ptree pt;
-  boost::property_tree::read_json(rig_config_path.c_str(), pt);
-
-  std::vector<CameraRig> camera_rigs;
-  for (const auto& rig_config : pt) {
-    CameraRig camera_rig;
-
-    std::vector<std::string> image_prefixes;
-    for (const auto& camera : rig_config.second.get_child("cameras")) {
-      const int camera_id = camera.second.get<int>("camera_id");
-      image_prefixes.push_back(camera.second.get<std::string>("image_prefix"));
-      camera_rig.AddCamera(camera_id, ComposeIdentityQuaternion(),
-                           Eigen::Vector3d(0, 0, 0));
-    }
-
-    camera_rig.SetRefCameraId(rig_config.second.get<int>("ref_camera_id"));
-
-    std::unordered_map<std::string, std::vector<image_t>> snapshots;
-    for (const auto image_id : reconstruction.RegImageIds()) {
-      const auto& image = reconstruction.Image(image_id);
-      for (const auto& image_prefix : image_prefixes) {
-        if (StringStartsWith(image.Name(), image_prefix)) {
-          const std::string image_suffix =
-              StringReplace(image.Name(), image_prefix, "");
-          snapshots[image_suffix].push_back(image_id);
-        }
-      }
-    }
-
-    for (const auto& snapshot : snapshots) {
-      bool has_ref_camera = false;
-      for (const auto image_id : snapshot.second) {
-        const auto& image = reconstruction.Image(image_id);
-        if (image.CameraId() == camera_rig.RefCameraId()) {
-          has_ref_camera = true;
-        }
-      }
-
-      if (has_ref_camera) {
-        camera_rig.AddSnapshot(snapshot.second);
-      }
-    }
-
-    camera_rig.Check(reconstruction);
-    camera_rig.ComputeRelativePoses(reconstruction);
-
-    camera_rigs.push_back(camera_rig);
-  }
-
-  return camera_rigs;
-}
-
-int main(int argc, char** argv) {
-  InitializeGlog(argv);
-
-  std::string input_path;
-  std::string output_path;
-  std::string rig_config_path;
-
-  OptionManager options;
-  options.AddRequiredOption("input_path", &input_path);
-  options.AddRequiredOption("output_path", &output_path);
-  options.AddRequiredOption("rig_config_path", &rig_config_path);
-  options.AddBundleAdjustmentOptions();
-  options.Parse(argc, argv);
-
-  Reconstruction reconstruction;
-  reconstruction.Read(input_path);
-
-  PrintHeading1("Camera rig configuration");
-
-  auto camera_rigs = ReadCameraRigConfig(rig_config_path, reconstruction);
-
-  BundleAdjustmentConfig config;
-  for (size_t i = 0; i < camera_rigs.size(); ++i) {
-    const auto& camera_rig = camera_rigs[i];
-    PrintHeading2(StringPrintf("Camera Rig %d", i + 1));
-    std::cout << StringPrintf("Cameras: %d", camera_rig.NumCameras())
-              << std::endl;
-    std::cout << StringPrintf("Snapshots: %d", camera_rig.NumSnapshots())
-              << std::endl;
-
-    // Add all registered images to the bundle adjustment configuration.
-    for (const auto image_id : reconstruction.RegImageIds()) {
-      config.AddImage(image_id);
-    }
-  }
-
-  PrintHeading1("Rig bundle adjustment");
-
-  BundleAdjustmentOptions ba_options = *options.bundle_adjustment;
-  ba_options.solver_options.minimizer_progress_to_stdout = true;
-  RigBundleAdjuster::Options rig_ba_options;
-  RigBundleAdjuster bundle_adjuster(ba_options, rig_ba_options, config);
-  CHECK(bundle_adjuster.Solve(&reconstruction, &camera_rigs));
-
-  reconstruction.Write(output_path);
-
-  return EXIT_SUCCESS;
-}
