+++ src/tools/CMakeLists.txt
@@ -1,4 +1,3 @@
 set(FOLDER_NAME "tools")
 
-# Example usage:
-# COLMAP_ADD_EXECUTABLE(executable executable.cc)
+# COLMAP_ADD_EXECUTABLE(example example.cc)
